Save on per-frame memory allocation.
In my app, over the course of a bit over a minute, 20Mb were allocated in handleModeTransitions,
even though every time that call was a NOOP. 18Mb of that alone was Function allocation. I believe
this could be refactored to be much more memory conscious, but this change makes an immediately and
obvious difference in my memory profiling and was super low-hanging. Clearly, this is memory that
would get GC'd eventually, and presumably AIR will force GC if iOS hands it a memory warning, but
that's a lot of junk that it's nice not to force the runtime to clean up.