Allow the texture source to be an embedded Bitmap.
Embedding my pngs with application/octet-stream in order to get them to be ByteArrays for the
loader.loadBytes() not only felt backwards, it was generating a content access error in my web
build. At the least, this should generate less garbage, and is also possibly faster.

This commit also fails the FontLoader if one of its child loaders fails.